+++ src/ScheduleParser.Core/AppInfo.cs
@@ -3,6 +3,6 @@ namespace ScheduleParser.Core
     public static class AppInfo
     {
         public const string ProductName = "Schedule Parser NUST MISIS";
-        public const string Version = "0.5.3";
+        public const string Version = "0.5.4";
     }
 }